refactor(ChangeDirectories): change packages for safe domain logic
- 도메인 안정성을 위해 Sections 생성 도메인 로직 관련 메서드 접근자들을 제한하기 위해 패키지 분리
