--- src/main/java/nextstep/subway/line/domain/AddSectionPolicy.java
@@ -1,5 +1,0 @@
-package nextstep.subway.line.domain;
-
-public interface AddSectionPolicy {
-    boolean addSection(final Section newSection);
-}
