+++ src/main/java/nextstep/subway/line/domain/sections/AddSectionPolicy.java
@@ -1,0 +1,7 @@
+package nextstep.subway.line.domain.sections;
+
+import nextstep.subway.line.domain.sections.Section;
+
+public interface AddSectionPolicy {
+    boolean addSection(final Section newSection);
+}
