+++ src/main/java/nextstep/subway/line/domain/sections/OriginalSectionCalculator.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.domain.sections;
 
 import nextstep.subway.line.domain.exceptions.NotExistSectionAddPolicy;
