+++ src/test/java/nextstep/subway/line/domain/LineTest.java
@@ -1,9 +1,7 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.line.domain.exceptions.EndUpStationNotFoundException;
-import nextstep.subway.line.domain.exceptions.InvalidSectionsActionException;
+import nextstep.subway.line.domain.sections.Section;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
