+++ src/test/java/nextstep/subway/line/domain/sections/ChangeOriginalAndAddSectionPolicyTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.domain.sections;
 
 import nextstep.subway.line.domain.exceptions.TargetSectionNotFoundException;
 import org.junit.jupiter.api.DisplayName;
