+++ src/test/java/nextstep/subway/line/domain/sections/OriginalSectionCalculatorTest.java
@@ -1,5 +1,7 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.domain.sections;
 
+import nextstep.subway.line.domain.sections.OriginalSectionCalculator;
+import nextstep.subway.line.domain.sections.Section;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
