+++ src/test/java/nextstep/subway/line/domain/sections/SimpleAddSectionPolicyTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.domain.sections;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
